ai_engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
httpx>=0.27.0
requests>=2.31.0
face_recognition==1.3.0
torch
dlib
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0
mlflow==2.8.1
boto3==1.28.0
